refactor: separate certificationlist show logic
